Update patch set 1
Patch Set 1:

I was under the same impression. I just know that the lib reports itself as supporting 1.1. Is the proper thing to do to report 1.0?

We're using the same lib for the Heroc (the one that shipped on our phone only supported 1.0); if you know of any specific benefits to reporting 1.0, I'll change it on the Heroc as well.

Patch-set: 1